requirements.txt
flask
flask-cors
Pillow
requests
python-dotenv
pyOpenSSL
cryptography
gunicorn
